adding ninja-keys
