+++ package.json
@@ -11,7 +11,6 @@
   "author": "Bridgetown Maintainers <maintainers@bridgetownrb.com>",
   "homepage": "https://www.bridgetownrb.com",
   "license": "MIT",
-  "private": false,
   "publishConfig": {
     "access": "public"
   },
@@ -27,6 +26,7 @@
   },
   "devDependencies": {
     "esbuild": "^0.15.13",
-    "glob": "^8.0.3"
+    "glob": "^8.0.3",
+    "konnors-ninja-keys": "^1.3.2"
   }
 }
